requirements.txt
flask
gunicorn
google-generativeai
python-dotenv
joblib
scikit-learn
textblob
transformers
#torch
#diffusers
#web3